fix Cypress tests (#7309)
