+++ e2e/cypress.json
@@ -13,7 +13,6 @@
         "dbClient": "postgres",
         "dbConnection": "postgres://mmuser:mostest@localhost/mattermost_test?sslmode=disable\u0026connect_timeout=10",
         "adminEmail": "sysadmin@sample.mattermost.com",
-        "ciBaseUrl": "",
         "enableApplitools": false,
         "enableVisualTest": false,
         "ldapServer": "localhost",
