+++ e2e/cypress/integration/messaging/header_spec.js
@@ -89,7 +89,7 @@ describe('Header', () => {
 
         // # Click the pin icon to open the pinned posts RHS
         cy.get('#channelHeaderPinButton').should('be.visible').click();
-        cy.get('#sidebar-right').should('be.visible').and('contain', 'Pinned posts');
+        cy.get('#sidebar-right').should('be.visible').and('contain', 'Pinned Posts');
 
         // # Verify that the Search term input box is still cleared and search term does not reappear when RHS opens
         cy.get('#searchBox').should('have.attr', 'value', '').and('be.empty');
