+++ e2e/cypress/integration/messaging/message_reply_too_long_spec.js
@@ -47,7 +47,7 @@ describe('Message Reply too long', () => {
         cy.get('#reply_textbox').type('{enter}');
 
         // * Check warning
-        cy.get('.post-error').should('be.visible').and('have.class', 'animation--highlight').and('have.text', `Your message is too long. Character count: ${replyTooLong.length}/${maxReplyLength}`);
+        cy.get('.post-error').should('be.visible').and('have.text', `Your message is too long. Character count: ${replyTooLong.length}/${maxReplyLength}`);
         cy.get('#reply_textbox').should('be.visible');
 
         // * Check last reply is the last valid one
