+++ e2e/cypress/integration/messaging/pinned_posts_spec.js
@@ -40,7 +40,7 @@ describe('Messaging', () => {
             cy.get('#channelHeaderPinButton').should('exist').click();
 
             // * RHS title displays as "Pinned Posts" and "[channel name]"
-            cy.get('#sidebar-right').should('be.visible').and('contain', 'Pinned posts').and('contain', `${testChannel.display_name}`);
+            cy.get('#sidebar-right').should('be.visible').and('contain', 'Pinned Posts').and('contain', `${testChannel.display_name}`);
 
             // * Pinned post appear in RHS
             cy.get(`#rhsPostMessageText_${postId}`).should('exist');
