+++ e2e/cypress/integration/messaging/post_pre_header_spec.js
@@ -135,7 +135,8 @@ describe('Post PreHeader', () => {
             cy.get('#searchContainer').should('be.visible').within(() => {
                 cy.get('.sidebar--right__title').
                     should('be.visible').
-                    and('have.text', 'Pinned postsTown Square');
+                    and('contain', 'Pinned Posts').
+                    and('contain', 'Town Square');
 
                 // * Check that the post pre-header is not shown for the pinned message in RHS
                 cy.findByTestId('search-item-container').within(() => {
