+++ e2e/cypress/integration/messaging/visual_verification_of_tooltips_on_top_nav_channel_icons_posts_spec.js
@@ -12,6 +12,7 @@
 
 describe('Messaging', () => {
     let testTeam;
+    let testChannelName;
 
     before(() => {
         // # Login as test user and visit the newly created test channel
@@ -26,7 +27,9 @@ describe('Messaging', () => {
             cy.apiLogin(user);
 
             // # Set up test channel with a long name
-            cy.apiCreateChannel(testTeam.id, 'channel-test', 'Public channel with a long name');
+            cy.apiCreateChannel(testTeam.id, 'channel-test', 'Public channel with a long name').then(({channel}) => {
+                testChannelName = channel.display_name;
+            });
             cy.visit(`/${testTeam.name}/channels/town-square`);
         });
     });
@@ -38,56 +41,61 @@ describe('Messaging', () => {
     });
 
     it('MM-T134 Visual verification of tooltips on top nav, channel icons, posts', () => {
-        // * Hover effect wraps around Date and New Messages lines
-        cy.get('#postListContent').find('.top').should('be.visible');
+        cy.findByRole('banner', {name: 'channel header region'}).should('be.visible').as('channelHeader');
 
-        // * Members tool-tip is present
-        cy.findAllByLabelText('members').should('be.visible').trigger('mouseover');
-        cy.get('div.tooltip-inner').should('be.visible').and('contain', 'Members');
-        cy.findAllByLabelText('members').should('be.visible').trigger('mouseout');
+        // * Members tooltip is present
+        openAndVerifyTooltip('members', 'Members');
 
-        // * Pinned post tool-tip is present
-        cy.findAllByLabelText('Pinned posts').should('be.visible').trigger('mouseover');
-        cy.get('div.tooltip-inner').should('be.visible').and('contain', 'Pinned posts');
-        cy.findAllByLabelText('Pinned posts').should('be.visible').trigger('mouseout');
+        // * Pinned post tooltip is present
+        openAndVerifyTooltip('Pinned posts', 'Pinned posts');
 
-        // * Saved posts tool-tip is present
-        cy.findAllByLabelText('Saved posts').should('be.visible').trigger('mouseover');
-        cy.get('div.tooltip-inner').should('be.visible').and('contain', 'Saved posts');
-        cy.findAllByLabelText('Saved posts').should('be.visible').trigger('mouseout');
+        // * Saved posts tooltip is present
+        openAndVerifyTooltip('Saved posts', 'Saved posts');
 
-        // * Add to favourites posts tool-tip is present - un checked
-        cy.findAllByLabelText('add to favorites').should('be.visible').trigger('mouseover');
-        cy.get('div.tooltip-inner').should('be.visible').and('contain', 'Add to Favorites');
-        cy.findAllByLabelText('add to favorites').should('be.visible').trigger('mouseout');
+        // * Add to favorites posts tooltip is present - un checked
+        openAndVerifyTooltip('add to favorites', 'Add to Favorites');
 
-        // * Add to favourites posts tool-tip is present - checked
-        cy.findAllByLabelText('add to favorites').click();
-        cy.findAllByLabelText('remove from favorites').should('be.visible').trigger('mouseover');
-        cy.get('div.tooltip-inner').should('be.visible').and('contain', 'Remove from Favorites');
-        cy.findAllByLabelText('remove from favorites').should('be.visible').trigger('mouseout');
+        // * Add to favorites posts tooltip is present - checked
+        cy.get('@channelHeader').findByRole('button', {name: 'add to favorites'}).should('be.visible').click();
+        openAndVerifyTooltip('remove from favorites', 'Remove from Favorites');
 
-        // * Unmute a channel tool-tip is present
-        cy.findAllByLabelText('dropdown icon').click();
-        cy.findAllByText('Mute Channel').click();
-        cy.findAllByLabelText('Muted Icon').should('be.visible').trigger('mouseover');
-        cy.get('div.tooltip-inner').should('be.visible').and('contain', 'Unmute');
-        cy.findAllByLabelText('Muted Icon').should('be.visible').trigger('mouseout');
+        // * Unmute a channel tooltip is present
+        cy.uiOpenChannelMenu('Mute Channel');
+        openAndVerifyTooltip('Muted Icon', 'Unmute');
 
         // * Download file tooltip is present
-        cy.findAllByLabelText('Upload files').attachFile('long_text_post.txt');
+        cy.findByLabelText('Upload files').attachFile('long_text_post.txt');
         cy.postMessage('test file upload');
-        cy.findAllByLabelText('download').should('be.visible').trigger('mouseover');
-        cy.get('div.tooltip-inner').should('be.visible').and('contain', 'Download');
-        cy.findAllByLabelText('download').should('be.visible').trigger('mouseout');
+        const downloadLink = () => cy.findByTestId('fileAttachmentList').should('be.visible').findByRole('link', {name: 'download', hidden: true});
+        downloadLink().trigger('mouseover');
+        verifyTooltip('Download');
+        downloadLink().trigger('mouseout');
 
         // * Long channel name (shown truncated on the LHS)
-        cy.get('.sidebar-item__name').contains('long name').should('be.visible').trigger('mouseover');
-        cy.get('div.tooltip-inner').should('be.visible').and('contain', 'Public channel with a long name');
-        cy.get('.sidebar-item__name').contains('long name').should('be.visible').trigger('mouseout');
+        cy.findByRole('application', {name: 'channel sidebar region'}).findByText(testChannelName).should('be.visible').as('longChannelAtSidebar');
+        cy.get('@longChannelAtSidebar').trigger('mouseover');
+        verifyTooltip(testChannelName);
+        cy.get('@longChannelAtSidebar').trigger('mouseout');
 
-        // * Check that the Demo plugin tool-tip is present
-        cy.get('#channel-header').find('.fa-plug').should('be.visible').trigger('mouseover');
-        cy.get('#pluginTooltip').should('be.visible').and('have.text', 'Demo Plugin');
+        // * Check that the Demo plugin tooltip is present
+        cy.get('@channelHeader').find('.fa-plug').should('be.visible').trigger('mouseover');
+        verifyTooltip('Demo Plugin');
     });
 });
+
+function verifyTooltip(text) {
+    cy.get('div.tooltip-inner').should('be.visible').and('have.text', text);
+}
+
+function openAndVerifyTooltip(name, text) {
+    const reName = RegExp(name, 'i');
+
+    // # Mouseover to the element
+    cy.get('@channelHeader').findByRole('button', {name: reName}).should('be.visible').trigger('mouseover');
+
+    // * Verify the tooltip
+    verifyTooltip(text);
+
+    // # Hide the tooltip
+    cy.get('@channelHeader').findByRole('button', {name: reName}).should('be.visible').trigger('mouseout');
+}
